test(firestore-send-email): add more tests
